added stles to stats
