+++ frontend/App.js
@@ -1,6 +1,6 @@
 import { View } from 'react-native';
 
-import StreakCounter from './components/StreakCounter';
+import StreakCounter from './components/streak-counter/StreakCounter';
 
 export default function App() {
   return (
