+++ frontend/components/streak-counter/streakCounter.style.js
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native"
+
+const styles = StyleSheet.create({
+  statContainer: {
+    marginTop: 40,
+    margin: 10,
+    borderWidth: 2,
+    borderColor: 'black',
+    borderRadius: 10,
+    padding: 10,
+  },
+  container: {
+    marginVertical: 10,
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    margin: 5,
+    color: 'black',
+  },
+})
+
+export default styles;
